Add WEB_FEATURES_MANIFEST.json generation
This will allow mapping of web features to web platform tests

Creates a new command: ./wpt web-features-manifest

It will create a file called WEB_FEATURES_MANIFEST.json inside of the
repo_root.

The format of the file is a simple dictionary.

The key is the web feature key from https://github.com/web-platform-dx/web-features/tree/main/feature-group-definitions

The value is a list of test paths that apply to the web feature key