compute pressure: Fix while loop to end earlier
On some testing bot, the while loop in testing seems
to take most of the time and the callback of the observer
is not invoked on time, which leads to timeout.

A step_timeout is also added to give a time to the event loop
to invoke the callback from the observer.

Bug: 1501324
Change-Id: I33e2253067e537e036b484e7e3fb62fe567346b9