Throw exceptions for malformed operations while layers are opened
Since layers are only filtered and composited into their parent bitmap
once the layer is closed, operations that access bitmap pixels are
malformed if there are layers open.

This can happen when reading the pixels from the canvas (for example
with `getImageData`) or directly overwriting pixels (with
`putImageData`).

With this change, exceptions are thrown in the missing cases for these
malformed operations. Promises also fail in the case of
`createImageBitmap` and `convertToBlob`.

The existing tests for unclosed layers had to be moved to manual, since for the offscreen and worker cases the generator uses methods that now throw exceptions.

Bug: 1396372
Change-Id: I3e187c9ae4f68135219a6178adf079616c470f9b