Tweak CloseWatcher user activation
This patch tweaks the CloseWatcher user activation checks so that
opening a dialog with user activation does not consume the user
activation to create the CloseWatcher. That way, when the user opens a
dialog element with user activation and then presses escape before
interacting with the page again, the cancel event still gets fired. This
is done in order to address the attached regression bug.

This patch also re-enables CloseWatcher now that the regression bug has
been addressed.

HTML spec PR: https://github.com/whatwg/html/pull/10048

Bug: 1512224
Change-Id: I6f39d538090a832130c40eb30f06041c7a31b741